Implemented content stabilization
ERM34767
[5.0.x]

Squashed commit of the following:

commit d571b59812a9f1401844edbb3d8255b29fa0ca0b
Merge: f34cdd5 026240a
Author: Daniel Riedmüller <daniel.riedmueller@hallowelt.com>
Date:   Wed Jan 29 13:33:50 2025 +0100

    Merge branch 'develop' into ERM34767


commit f34cdd5aa5cc22faa949a7831a4f7f037a872b76
Author: Daniel Riedmüller <daniel.riedmueller@hallowelt.com>
Date:   Wed Jan 29 13:33:45 2025 +0100

    Use tabs instead of spaces

    ERM34767


commit 0a5ce3133ef71fa4a22eeb6b542079f4fa09b165
Author: Daniel Riedmüller <daniel.riedmueller@hallowelt.com>
Date:   Wed Jan 29 13:23:02 2025 +0100

    Fix check for existing content slot

    ERM34767

commit f85a7536eac50baa6db198d9c81362cc4efe4cff
Merge: 3748312 2979e00
Author: Daniel Riedmüller <daniel.riedmueller@hallowelt.com>
Date:   Wed Jan 29 13:11:33 2025 +0100

    Merge branch 'master' into ERM34767

    # Conflicts:
    #	resources/templates/CpdContainer.mustache
    #	src/Api/LoadCpdDiagram.php
    #	src/Data/OrphanedDescriptionPages/SecondaryDataProvider.php
    #	src/HookHandler/BpmnTag.php
    #	src/Util/CpdDescriptionPageUtil.php
    #	src/Util/CpdDiagramPageUtil.php
    #	src/Util/CpdSaveDescriptionPagesUtil.php

commit 3748312390c16dea065787e58776e994849fc860
Author: d.savuljesku <d.savuljesku@gmail.com>
Date:   Wed Jan 29 08:52:50 2025 +0100

    Fixes


commit 2593527ab1941e00283adf0390726a25d33ec8da
Author: d.savuljesku <d.savuljesku@gmail.com>
Date:   Tue Jan 28 14:06:46 2025 +0100

    Fix null check


commit d5cbbe01a5157d7e4b1780de0b11b2d86945b772
Author: d.savuljesku <d.savuljesku@gmail.com>
Date:   Tue Jan 28 13:50:24 2025 +0100

    Fix CI


commit 75a0180bb5ee060cd1a4f63e5eb97481eae17b59
Author: d.savuljesku <d.savuljesku@gmail.com>
Date:   Tue Jan 28 12:48:46 2025 +0100

    CS integration


commit 1358b1d61737a150aa284a81b590a3b846342890
Author: Daniel Riedmüller <daniel.riedmueller@hallowelt.com>
Date:   Tue Jan 28 10:11:01 2025 +0100

    add missing i18n key

    ERM34767

commit ff3b6f8e471476b17d4fbed601e92403aadd10f0
Merge: 7f3f2c6 afa7c92
Author: Daniel Riedmüller <daniel.riedmueller@hallowelt.com>
Date:   Fri Jan 24 15:08:07 2025 +0100

    Merge branch 'develop' into ERM34767

commit 7f3f2c63c1fb876ec61b2e3750c8ea1b35826a0e
Merge: 098a96f e3eca9a
Author: Daniel Riedmüller <daniel.riedmueller@hallowelt.com>
Date:   Fri Jan 24 09:00:12 2025 +0100

    Merge branch 'master' into ERM34767

    # Conflicts:
    #	resources/templates/CpdContainer.mustache
    #	src/Api/LoadCpdDiagram.php
    #	src/Data/OrphanedDescriptionPages/SecondaryDataProvider.php
    #	src/HookHandler/BpmnTag.php
    #	src/Util/CpdDescriptionPageUtil.php
    #	src/Util/CpdDiagramPageUtil.php
    #	src/Util/CpdSaveDescriptionPagesUtil.php

commit 098a96fb0668bc322cb99e6e92b60365ad10c9aa
Merge: ec8d5c6 bcc06b9
Author: Daniel Riedmüller <daniel.riedmueller@hallowelt.com>
Date:   Fri Jan 17 14:20:44 2025 +0100

    Merge branch 'master' into ERM34767


    # Conflicts:
    #	resources/js/dist/cpd.viewer.bundle.js

commit ec8d5c6243d51f8f48e2a858a586178c4e4b9264
Merge: d1df9a0 060c0f5
Author: Daniel Riedmüller <daniel.riedmueller@hallowelt.com>
Date:   Fri Jan 17 13:00:36 2025 +0100

    Merge branch 'master' into ERM34767


    # Conflicts:
    #	resources/js/dist/cpd.viewer.bundle.js

commit d1df9a0ca061a943fba411b2addd76a12af614a4
Merge: 48a67f9 d867662
Author: Daniel Riedmüller <daniel.riedmueller@hallowelt.com>
Date:   Fri Jan 17 12:24:08 2025 +0100

    Merge branch 'master' into ERM34767


    # Conflicts:
    #	resources/js/dist/cpd.viewer.bundle.js

commit 48a67f99935b25e86580b2d08b1ba8cc23b1f12a
Merge: 95dbfff df71cee
Author: Daniel Riedmüller <daniel.riedmueller@hallowelt.com>
Date:   Thu Jan 16 11:51:22 2025 +0100

    Merge branch 'master' into ERM34767


    # Conflicts:
    #	resources/js/dist/cpd.viewer.bundle.js

commit 95dbfff126cfba4aee49d7d53b5eb798de57ae01
Author: Daniel Riedmüller <daniel.riedmueller@hallowelt.com>
Date:   Wed Jan 15 13:56:33 2025 +0100

    apply content stabilization

    ERM34767


commit 85e9e77d0de0eb1c3178727a585496cea0a29eb2
Merge: be35945 4ae9f33
Author: Daniel Riedmüller <daniel.riedmueller@hallowelt.com>
Date:   Wed Jan 15 09:54:33 2025 +0100

    Merge branch 'master' into ERM34767


commit be359458db163aaac7532b3a300e1726b645254b
Author: Daniel Riedmüller <daniel.riedmueller@hallowelt.com>
Date:   Wed Jan 15 09:54:18 2025 +0100

    minor fixes


commit 2e00a6d41901f7d0808d813b474ec69936fd9d91
Author: Daniel Riedmueller <daniel.riedmueller@hallowelt.com>
Date:   Tue Jan 14 15:50:08 2025 +0100

    show only stabilized revisions in orphaned description pages list

    ERM34767

    [5.0.x]

commit d0d8e89d95edc2614e42598877eb4edbf5e7d296
Author: Daniel Riedmüller <daniel.riedmueller@hallowelt.com>
Date:   Mon Jan 13 12:03:58 2025 +0100

    save rev along orphaned pages

    ERM34767


commit 7074dbb1a531578f094e159d1df543b1b99bdf05
Merge: 1c46f3a d55d15e
Author: Daniel Riedmüller <daniel.riedmueller@hallowelt.com>
Date:   Fri Jan 10 14:39:29 2025 +0100

    Merge branch 'master' into ERM34767


    # Conflicts:
    #	resources/js/dist/cpd.viewer.bundle.js

commit 1c46f3a858377d0d7dfe9f4deab9dbfb6721b4b7
Author: Daniel Riedmüller <daniel.riedmueller@hallowelt.com>
Date:   Fri Jan 10 14:21:49 2025 +0100

    content stabilization for process namespace

    ERM34767


commit e5b982bb9074832010c94891ab2199a115e4a440
Merge: ae7623b 92c7e8f
Author: Daniel Riedmüller <daniel.riedmueller@hallowelt.com>
Date:   Fri Jan 10 12:03:40 2025 +0100

    Merge branch 'master' into ERM34767


commit ae7623b5008426886b536d94ea043e6f9f088be8
Author: Daniel Riedmüller <daniel.riedmueller@hallowelt.com>
Date:   Fri Jan 10 08:54:55 2025 +0100

    use revision on xml load

    ERM34767

+++ src/Util/CpdDescriptionPageUtil.php
@@ -7,36 +7,30 @@ use CognitiveProcessDesigner\CpdElement;
 use CognitiveProcessDesigner\Exceptions\CpdSaveException;
 use MediaWiki\Config\Config;
 use MediaWiki\Content\Content;
+use MediaWiki\Extension\ContentStabilization\StabilizationLookup;
 use MediaWiki\Page\PageStore;
 use MediaWiki\Page\WikiPageFactory;
 use MediaWiki\Title\Title;
 use Wikimedia\Rdbms\ILoadBalancer;
 
 class CpdDescriptionPageUtil {
-	/**
-	 * @var PageStore
-	 */
+	/** @var PageStore */
 	private PageStore $pageStore;
 
-	/**
-	 * @var ILoadBalancer
-	 */
+	/** @var ILoadBalancer */
 	private ILoadBalancer $loadBalancer;
 
-	/**
-	 * @var Config
-	 */
+	/** @var Config */
 	private Config $config;
 
-	/**
-	 * @var WikiPageFactory
-	 */
+	/** @var WikiPageFactory */
 	private WikiPageFactory $wikiPageFactory;
 
-	/**
-	 * @var CpdElementConnectionUtil
-	 */
+	/** @var CpdElementConnectionUtil */
 	private CpdElementConnectionUtil $connectionUtil;
+
+	/** @var StabilizationLookup */
+	private StabilizationLookup $lookup;
 
 	/**
 	 * @param PageStore $pageStore
@@ -44,19 +38,22 @@ class CpdDescriptionPageUtil {
 	 * @param WikiPageFactory $wikiPageFactory
 	 * @param Config $config
 	 * @param CpdElementConnectionUtil $connectionUtil
+	 * @param StabilizationLookup $lookup
 	 */
 	public function __construct(
 		PageStore $pageStore,
 		ILoadBalancer $loadBalancer,
 		WikiPageFactory $wikiPageFactory,
 		Config $config,
-		CpdElementConnectionUtil $connectionUtil
+		CpdElementConnectionUtil $connectionUtil,
+		StabilizationLookup $lookup
 	) {
 		$this->pageStore = $pageStore;
 		$this->loadBalancer = $loadBalancer;
 		$this->config = $config;
 		$this->wikiPageFactory = $wikiPageFactory;
 		$this->connectionUtil = $connectionUtil;
+		$this->lookup = $lookup;
 	}
 
 	/**
@@ -135,18 +132,21 @@ class CpdDescriptionPageUtil {
 	/**
 	 * @param CpdElement[] $elements
 	 * @param string $process
+	 * @param int $revision
 	 *
 	 * @return void
 	 */
-	public function updateOrphanedDescriptionPages( array $elements, string $process ): void {
+	public function updateOrphanedDescriptionPages( array $elements, string $process, int $revision ): void {
 		$dbw = $this->loadBalancer->getConnection( DB_PRIMARY );
 
-		// Clear orphaned pages rows from this process
-		$dbw->delete(
-			'cpd_orphaned_description_pages',
-			[ 'process' => $process ],
-			__METHOD__
-		);
+		if ( !$this->isStabilizationEnabled() ) {
+			// Clear orphaned pages rows from this process
+			$dbw->delete(
+				'cpd_orphaned_description_pages',
+				[ 'process' => $process ],
+				__METHOD__
+			);
+		}
 
 		$orphanedPages = [];
 		$existingPages = array_map( fn ( Title $title ) => $title->getPrefixedDBkey(),
@@ -166,13 +166,37 @@ class CpdDescriptionPageUtil {
 			return;
 		}
 
-		// Insert orphaned pages
 		$dbw->insert(
 			'cpd_orphaned_description_pages',
-			array_map( fn ( string $page ) => [
-				'process' => $process,
-				'page_title' => $page
-			], $orphanedPages ),
+		array_map( fn( string $page ) => [
+			'process' => $process,
+			'process_rev' => $revision,
+			'page_title' => $page
+		], $orphanedPages ), __METHOD__, [ 'IGNORE' ]
+		);
+	}
+
+	/**
+	 * Removes all orphaned description pages for the given process
+	 * except the ones that are in the given revision when given.
+	 * Runs when a new a process is stabilized.
+	 *
+	 * @param string $process
+	 * @param int|null $revision
+	 *
+	 * @return void
+	 */
+	public function cleanUpOrphanedDescriptionPages( string $process, int $revision = null ): void {
+		$dbw = $this->loadBalancer->getConnectionRef( DB_PRIMARY );
+
+		$conds = [ 'process' => $process ];
+		if ( $revision ) {
+			$conds[] = 'process_rev != ' . $revision;
+		}
+
+		$dbw->delete(
+			'cpd_orphaned_description_pages',
+			$conds,
 			__METHOD__
 		);
 	}
@@ -185,5 +209,11 @@ class CpdDescriptionPageUtil {
 	 */
 	public function updateElementConnections( array $elements, string $process ): void {
 		$this->connectionUtil->updateElementConnections( $elements, $process );
+	}
+
+	private function isStabilizationEnabled(): bool {
+		$dummyPage = Title::newFromText( 'Dummy', NS_PROCESS );
+
+		return $this->lookup->isStabilizationEnabled( $dummyPage );
 	}
 }
